+++ Source/AIURLShortenerPlugin.h
@@ -20,6 +20,10 @@ typedef enum { AITinyURL = 0, AIisgd, AIMetamark, AINumberOfShortenLinkServices 
 
 @interface AIURLShortenerPlugin : AIPlugin <NSMenuDelegate> {
 	AIShortenLinkService shortener;
+	NSMenuItem *menuItem_shortenLink;
+	NSMenuItem *menuItem_shortenLinkContext;
+	NSMenu *shortenerSubMenu;
+	NSMenu *shortenerSubMenuContext;
 }
 
 @end